[v0.6.0] Pipeline configuration schema (#120)
* [Flink][Storm] Added tuple id for debug

* [Conf] Added filtering property

* [Conf] Added pipeline definition

* [Conf] Added filtering property to pipeline

* added pipeline configurations

* fixed null bug

* fixed source naming

* updated configuration namings; versioning

* updated output on non-windowed tasks

* fixed windowed task; readded heron

* [0.6] Bump version

+++ benchmarks/utils/src/main/java/fr/unice/yamb/utils/configuration/schema/StormConfigSchema.java
@@ -8,7 +8,7 @@ public class StormConfigSchema extends ConfigSchema {
     private int workers = 1;
     private int maxSpoutPending = 5000;
     private StormDeployment deployment = StormDeployment.local;
-    private float debugFrequency = 0;
+    private double debugFrequency = 0.0;
 
     public int getWorkers() {
         return workers;
@@ -34,11 +34,11 @@ public class StormConfigSchema extends ConfigSchema {
         this.maxSpoutPending = maxSpoutPending;
     }
 
-    public float getDebugFrequency() {
+    public double getDebugFrequency() {
         return debugFrequency;
     }
 
-    public void setDebugFrequency(float debugFrequency) {
+    public void setDebugFrequency(double debugFrequency) {
         this.debugFrequency = debugFrequency;
     }
 }
